feat(db): Integrate changeset cache into OverlayStateProviderFactory
