+++ crates/engine/service/src/service.rs
@@ -144,10 +144,7 @@ mod tests {
     use alloy_eips::merge::EPOCH_SLOTS;
     use reth_chainspec::{ChainSpecBuilder, MAINNET};
     use reth_engine_primitives::{BeaconEngineMessage, NoopInvalidBlockHook};
-    use reth_engine_tree::{
-        cache::changeset_cache::ChangesetCache, test_utils::TestPipelineBuilder,
-        tree::BasicEngineValidator,
-    };
+    use reth_engine_tree::{test_utils::TestPipelineBuilder, tree::BasicEngineValidator};
     use reth_ethereum_consensus::EthBeaconConsensus;
     use reth_ethereum_engine_primitives::EthEngineTypes;
     use reth_evm_ethereum::EthEvmConfig;
@@ -160,7 +157,8 @@ mod tests {
     };
     use reth_prune::Pruner;
     use reth_tasks::TokioTaskExecutor;
-    use std::sync::{Arc, RwLock};
+    use reth_trie_db::changesets::ChangesetCacheHandle;
+    use std::sync::Arc;
     use tokio::sync::{mpsc::unbounded_channel, watch};
     use tokio_stream::wrappers::UnboundedReceiverStream;
 
@@ -193,7 +191,7 @@ mod tests {
         let evm_config = EthEvmConfig::new(chain_spec.clone());
 
         // Cache size is 2 epochs (64 blocks) to cover the finalization window
-        let changeset_cache = Arc::new(RwLock::new(ChangesetCache::new(EPOCH_SLOTS * 2)));
+        let changeset_cache = ChangesetCacheHandle::new(EPOCH_SLOTS * 2);
 
         let engine_validator = BasicEngineValidator::new(
             blockchain_db.clone(),
