+++ crates/engine/tree/Cargo.toml
@@ -27,6 +27,7 @@ reth-primitives-traits.workspace = true
 reth-ethereum-primitives.workspace = true
 reth-provider.workspace = true
 reth-prune.workspace = true
+reth-storage-api.workspace = true
 reth-revm.workspace = true
 reth-stages-api.workspace = true
 reth-tasks.workspace = true
@@ -35,6 +36,7 @@ reth-trie-sparse = { workspace = true, features = ["std", "metrics"] }
 reth-trie-sparse-parallel = { workspace = true, features = ["std"] }
 reth-trie-common.workspace = true
 reth-trie.workspace = true
+reth-trie-db.workspace = true
 
 # alloy
 alloy-evm.workspace = true
