+++ crates/trie/db/Cargo.toml
@@ -20,9 +20,13 @@ reth-trie.workspace = true
 reth-trie-common.workspace = true
 reth-storage-api = { workspace = true, features = ["db-api"] }
 reth-storage-errors.workspace = true
+reth-metrics = { workspace = true, optional = true }
 
 # alloy
 alloy-primitives.workspace = true
+
+# metrics
+metrics = { workspace = true, optional = true }
 
 # tracing
 tracing.workspace = true
@@ -51,7 +55,7 @@ serde_json.workspace = true
 similar-asserts.workspace = true
 
 [features]
-metrics = ["reth-trie/metrics"]
+metrics = ["reth-trie/metrics", "dep:reth-metrics", "dep:metrics"]
 serde = [
     "similar-asserts/serde",
     "alloy-consensus/serde",
